+++ lib/auth/trustedcluster_test.go
@@ -42,7 +42,6 @@ import (
 	"github.com/gravitational/teleport/lib/modules"
 	"github.com/gravitational/teleport/lib/modules/modulestest"
 	"github.com/gravitational/teleport/lib/services"
-	"github.com/gravitational/teleport/lib/services/suite"
 )
 
 func TestRemoteClusterStatus(t *testing.T) {
@@ -250,8 +249,8 @@ func TestValidateTrustedCluster(t *testing.T) {
 		_, err = a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token: validToken,
 			CAs: []types.CertAuthority{
-				suite.NewTestCA(types.HostCA, "rc1"),
-				suite.NewTestCA(types.HostCA, "rc2"),
+				authtest.NewTestCA(types.HostCA, "rc1"),
+				authtest.NewTestCA(types.HostCA, "rc2"),
 			},
 		})
 		require.Error(t, err)
@@ -262,7 +261,7 @@ func TestValidateTrustedCluster(t *testing.T) {
 		_, err = a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token: validToken,
 			CAs: []types.CertAuthority{
-				suite.NewTestCA(types.UserCA, "rc3"),
+				authtest.NewTestCA(types.UserCA, "rc3"),
 			},
 		})
 		require.Error(t, err)
@@ -273,7 +272,7 @@ func TestValidateTrustedCluster(t *testing.T) {
 		_, err = a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token: validToken,
 			CAs: []types.CertAuthority{
-				suite.NewTestCA(types.HostCA, localClusterName),
+				authtest.NewTestCA(types.HostCA, localClusterName),
 			},
 		})
 		require.Error(t, err)
@@ -292,7 +291,7 @@ func TestValidateTrustedCluster(t *testing.T) {
 		_, err = a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token: validToken,
 			CAs: []types.CertAuthority{
-				suite.NewTestCA(types.HostCA, trustedCluster.GetName()),
+				authtest.NewTestCA(types.HostCA, trustedCluster.GetName()),
 			},
 		})
 		require.Error(t, err)
@@ -300,7 +299,7 @@ func TestValidateTrustedCluster(t *testing.T) {
 	})
 
 	t.Run("all CAs are returned when v10+", func(t *testing.T) {
-		leafClusterCA := types.CertAuthority(suite.NewTestCA(types.HostCA, "leafcluster-1"))
+		leafClusterCA := types.CertAuthority(authtest.NewTestCA(types.HostCA, "leafcluster-1"))
 		resp, err := a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token:           validToken,
 			CAs:             []types.CertAuthority{leafClusterCA},
@@ -375,7 +374,7 @@ func TestValidateTrustedCluster(t *testing.T) {
 	})
 
 	t.Run("Host User and Database CA are returned by default", func(t *testing.T) {
-		leafClusterCA := types.CertAuthority(suite.NewTestCA(types.HostCA, "leafcluster-2"))
+		leafClusterCA := types.CertAuthority(authtest.NewTestCA(types.HostCA, "leafcluster-2"))
 		resp, err := a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token:           validToken,
 			CAs:             []types.CertAuthority{leafClusterCA},
@@ -409,7 +408,7 @@ func TestValidateTrustedCluster(t *testing.T) {
 		_, err = a.ValidateTrustedCluster(ctx, &authclient.ValidateTrustedClusterRequest{
 			Token: validToken,
 			CAs: []types.CertAuthority{
-				suite.NewTestCAWithConfig(suite.TestCAConfig{
+				authtest.NewTestCAWithConfig(authtest.TestCAConfig{
 					Type:                types.HostCA,
 					ClusterName:         "remoteCluster",
 					SubjectOrganization: "commonName",
@@ -499,7 +498,7 @@ func TestUpsertTrustedCluster(t *testing.T) {
 	trustedCluster, err := types.NewTrustedCluster("trustedcluster", trustedClusterSpec)
 	require.NoError(t, err)
 
-	ca := suite.NewTestCA(types.UserCA, "trustedcluster")
+	ca := authtest.NewTestCA(types.UserCA, "trustedcluster")
 
 	auth.ConfigureCAsForTrustedCluster(trustedCluster, []types.CertAuthority{ca})
 
@@ -638,7 +637,7 @@ func TestUpdateTrustedCluster(t *testing.T) {
 	trustedCluster, err := types.NewTrustedCluster(testClusterName, trustedClusterSpec)
 	require.NoError(t, err)
 
-	ca := suite.NewTestCA(types.UserCA, testClusterName)
+	ca := authtest.NewTestCA(types.UserCA, testClusterName)
 
 	auth.ConfigureCAsForTrustedCluster(trustedCluster, []types.CertAuthority{ca})
