+++ lib/services/local/users.go
@@ -31,7 +31,6 @@ import (
 	"sort"
 	"strings"
 	"sync"
-	"testing"
 	"time"
 
 	"github.com/gogo/protobuf/jsonpb"
@@ -90,11 +89,6 @@ func NewIdentityService(backend backend.Backend) (*IdentityService, error) {
 // used in tests. It will use weaker cryptography to minimize the time it takes
 // to perform flakiness tests and decrease the probability of timeouts.
 func NewTestIdentityService(backend backend.Backend) (*IdentityService, error) {
-	if !testing.Testing() {
-		// Don't allow using weak cryptography in production.
-		panic("Attempted to create a test identity service outside of a test")
-	}
-
 	s, err := NewIdentityService(backend)
 	if err != nil {
 		return nil, trace.Wrap(err)
